+++ FoodOrdering.Presentation/Controllers/DishController.cs
@@ -4,6 +4,10 @@ using System.Threading.Tasks;
 
 using FoodOrdering.Application.DishService;
 using FoodOrdering.Domain.Entities;
+using FoodOrdering.Presentation.ViewModels;
+using AutoMapper;
+using FoodOrdering.Application.Infrastructure;
+using FoodOrdering.Application.DTO.Dish;
 
 // For more information on enabling Web API for empty projects, visit https://go.microsoft.com/fwlink/?LinkID=397860
 
@@ -11,7 +15,7 @@ namespace FoodOrdering.Presentation.Controllers
 {
     [Route("api/")]
     [ApiController]
-    public class DishController : ControllerBase
+    public class DishController : Controller
     {
         private readonly IDishService _service;
 
@@ -24,11 +28,7 @@ namespace FoodOrdering.Presentation.Controllers
         public async Task<IActionResult> GetAllDishes()
         {
             var result = await _service.GetAllDishesAsync();
-            if (result is not null)
-            {
-                return Ok(result);
-            }
-            return BadRequest("No dishes found");
+            return Ok(result);
         }
 
         [HttpGet(nameof(GetDish) + "/{id}")]
@@ -42,34 +42,67 @@ namespace FoodOrdering.Presentation.Controllers
             return BadRequest("Dish not found");
         }
 
-        [HttpPost(nameof(CreateDish))]
-        public async Task<IActionResult> CreateDish(Dish dish)
+        [HttpPost(nameof(AddDish))]
+        public async Task<IActionResult> AddDish(AddDishViewModel model)
         {
-            await _service.CreateDishAsync(dish);
+            DishDTO dishDto = new()
+            {
+                Name = model.Name,
+                Ingredients = model.Ingredients,
+                Price = model.Price,
+                Proteins = model.Proteins,
+                Fats = model.Fats,
+                Carbohydrates = model.Carbohydrates,
+                Calories = model.Calories,
+                Weight = model.Weight
+            };
+
+            await _service.CreateDishAsync(dishDto);
             return Ok("Dish has been created");
         }
 
-        [HttpPut(nameof(UpdateDish)+"/{id}")]
-        public async Task<IActionResult> UpdateDish(Guid id, Dish dish)
+        [HttpPut(nameof(EditDish)+"/{id}")]
+        public async Task<IActionResult> EditDish(Guid id, EditDishViewModel model)
         {
-            if (id != dish.Id)
+            if (ModelState.IsValid)
             {
-                return BadRequest();
-            }
+                try { 
+                    DishDTO dishDto = new()
+                    {
+                        Id = id,
+                        Name = model.Name,
+                        Ingredients = model.Ingredients,
+                        Price = model.Price,
+                        Proteins = model.Proteins,
+                        Fats = model.Fats,
+                        Carbohydrates = model.Carbohydrates,
+                        Calories = model.Calories,
+                        Weight = model.Weight
+                    };
 
-            await _service.UpdateDishAsync(dish);
-            return Ok("Dish has been updated");
+                    await _service.UpdateDishAsync(dishDto);
+                    return Ok("Dish has been updated");
+                }
+                catch(ValidationException ex)
+                {
+                    return NotFound(ex.Message);
+                }
+            }
+            else
+            {
+                return View(model);
+            }
         }
 
         [HttpDelete(nameof(DeleteDish) + "/{id}")]
         public async Task<IActionResult> DeleteDish(Guid id)
         {
-            try
+            if (await _service.DishExists(id))
             {
                 await _service.DeleteDishAsync(id);
                 return Ok("Dish has been deleted");
             }
-            catch
+            else
             {
                 return NotFound("Dish not found");
             }
